refactor test directory
